+++ src/commonTest/kotlin/com/insanusmokrassar/krontab/utils/FailJob.kt
@@ -1,8 +1,0 @@
-package com.insanusmokrassar.krontab.utils
-
-import kotlinx.coroutines.*
-
-fun CoroutineScope.createFailJob(forTimeMillis: Long) = launch {
-    delay(forTimeMillis)
-    throw IllegalStateException("This job must be completed at before this exception happen")
-}