+++ src/commonTest/kotlin/dev/inmo/krontab/utils/FailJob.kt
@@ -1,0 +1,8 @@
+package dev.inmo.krontab.utils
+
+import kotlinx.coroutines.*
+
+fun CoroutineScope.createFailJob(forTimeMillis: Long) = launch {
+    delay(forTimeMillis)
+    throw IllegalStateException("This job must be completed at before this exception happen")
+}