+++ src/commonTest/kotlin/dev/inmo/krontab/utils/RunTest.kt
@@ -1,0 +1,8 @@
+package dev.inmo.krontab.utils
+
+import kotlinx.coroutines.CoroutineScope
+
+/**
+ * Workaround to use suspending functions in unit tests
+ */
+expect fun runTest(block: suspend (scope : CoroutineScope) -> Unit)