+++ src/commonTest/kotlin/dev/inmo/krontab/utils/SchedulerFlow.kt
@@ -1,0 +1,68 @@
+package dev.inmo.krontab.utils
+
+import dev.inmo.krontab.builder.buildSchedule
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.takeWhile
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+@ExperimentalCoroutinesApi
+@FlowPreview
+class SchedulerFlowTests {
+    @Test
+    fun testThatFlowIsCorrectlyWorkEverySecond() {
+        val kronScheduler = buildSchedule {
+            seconds {
+                0 every 1
+            }
+        }
+
+        val flow = kronScheduler.asFlow()
+
+        runTest {
+            val mustBeCollected = 10
+            var collected = 0
+            flow.takeWhile {
+                collected < mustBeCollected
+            }.collect {
+                collected++
+            }
+            assertEquals(mustBeCollected, collected)
+        }
+    }
+
+    @Test
+    fun testThatFlowIsCorrectlyWorkEverySecondWithMuchOfEmitters() {
+        val kronScheduler = buildSchedule {
+            seconds {
+                0 every 1
+            }
+        }
+
+        val flow = kronScheduler.asFlow()
+
+        runTest {
+            val testsCount = 10
+            val failJob = it.createFailJob((testsCount * 2) * 1000L)
+            val mustBeCollected = 10
+            val answers = (0 until testsCount).map { _ ->
+                it.async {
+                    var collected = 0
+                    flow.takeWhile {
+                        collected < mustBeCollected
+                    }.collect {
+                        collected++
+                    }
+                    collected
+                }
+            }.awaitAll()
+
+            failJob.cancel()
+
+            answers.forEach {
+                assertEquals(mustBeCollected, it)
+            }
+        }
+    }
+}